Extract dart_type.dart
